Shorten the text in the nav menu
Shortening the text to fix a small range of browser window sizes in which
the user menu and the nav menu fight for space, forcing the user menu to
float down resulting in being hidden by the announcements div.

This is a bit of a punt to avoid hunting down the CSS that rules these
widths.

Signed-off-by: Robb Kidd <rkidd@chef.io>