add an abort during initialization if only a partial S3 config is supplied
With the upgrade of paperclip and the aws-sdk, S3_REGION is now required
else calls can't be made to the S3 API. Because there are existing installs
that probably lack this environment variable until a human can add it to
the configuration, we feel it is better to abort starting the web app than
to start and store cookbooks on disk unbeknownst to the admins and user.

Moved the logic of whether Supermarket is configured for S3 to a library
class to make testing the cases clearer. Initializers are wicked and
unforgiving; keep them simple.

Signed-off-by: Robb Kidd <rkidd@chef.io>